pyproject.toml
[tool.poetry]
name = "geolocator"
version = "0.1.0"
description = "A FastAPI application for geolocation using Nominatim"
authors = ["Mable Zhang <mablezhang2003@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.88.0"
httpx = "^0.24.0"
geopy = "^2.4.1"
uvicorn = "^0.35.0"
psycopg2 = "^2.9.3"
SQLAlchemy = "^2.0.0"
python-dotenv = "^1.1.1"
asyncpg = "^0.30.0"
requests = "^2.32.4"
pandas = "^2.3.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
